fix: convert to TS and fix type issues
refactor: migrate to typescript and fix type issues

fix: fix type issues on topology
